+++ build.gradle
@@ -67,11 +67,11 @@ dependencies {
     implementation 'com.googlecode.aviator:aviator:5.4.3'
     implementation 'org.springframework.boot:spring-boot-starter-mail'
     implementation "com.github.ben-manes.caffeine:caffeine:3.2.2"
-    implementation 'io.github.resilience4j:resilience4j-core:2.2.0'
-    implementation 'io.github.resilience4j:resilience4j-retry:2.2.0'
-    implementation 'io.github.resilience4j:resilience4j-timelimiter:2.2.0'
-    implementation 'io.github.resilience4j:resilience4j-ratelimiter:2.2.0'
-    implementation 'io.github.resilience4j:resilience4j-bulkhead:2.2.0'
+    implementation 'io.github.resilience4j:resilience4j-core:2.3.0'
+    implementation 'io.github.resilience4j:resilience4j-retry:2.3.0'
+    implementation 'io.github.resilience4j:resilience4j-timelimiter:2.3.0'
+    implementation 'io.github.resilience4j:resilience4j-ratelimiter:2.3.0'
+    implementation 'io.github.resilience4j:resilience4j-bulkhead:2.3.0'
 
     implementation 'net.datafaker:datafaker:2.5.0'
